requirements.txt
streamlit==1.38.0
pandas==2.2.2
qrcode==7.4.2
Pillow==10.4.0
xlsxwriter==3.2.0
python-dotenv==1.0.1